# api_geteway/api/docs/swagger.yaml
definitions:
...
  models.ChangePassword:
    properties:
      newPassword:
        type: string
      oldPassword:
        type: string
    type: object
...
  models.CreateStaff:
    properties:
      age:
        type: integer
      balance:
        type: number
      birth_date:
        type: string
      branch_id:
        type: string
      loging:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type:
        type: string
      tarif_id:
        type: string
    type: object
...
  models.Staff:
    properties:
      age:
        type: integer
      balance:
        type: number
      birth_date:
        type: string
      branch_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      loging:
        type: string
      name:
        type: string
      password:
        type: string
      staff_type:
        type: string
      tarif_id:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResp:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /change-password/{id}:
    put:
      consumes:
      - application/json
      description: change staff's password
      parameters:
      - description: Staff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: data of password
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePassword'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Change Password
      tags:
      - PASSWORD
...
  /staffs:
    get:
      consumes:
      - application/json
      description: get Staff
      parameters:
      - default: 10
        description: limit for response
        in: query
        name: limit
        required: true
        type: integer
      - default: 1
        description: page of req
        in: query
        name: page
        required: true
        type: integer
      - description: filter by name
        in: query
        name: name
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Staff'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: List Staffs
      tags:
      - STAFF
    post:
      consumes:
      - application/json
      description: Create Stsff
      parameters:
      - description: staff data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Create Staff
      tags:
      - STAFF
  /staffs/{id}:
    delete:
      consumes:
      - application/json
      description: delete staff by Id
      parameters:
      - description: Staff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Delete By Id
      tags:
      - STAFF
    get:
      consumes:
      - application/json
      description: get staff by ID
      parameters:
      - description: Staff ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Staff'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Get By Id
      tags:
      - STAFF
    put:
      consumes:
      - application/json
      description: api for update staffs
      parameters:
      - description: id of staff
        in: path
        name: id
        required: true
        type: string
      - description: data of staff
        in: body
        name: staff
        required: true
        schema:
          $ref: '#/definitions/models.CreateStaff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResp'
      summary: Update Staff
      tags:
      - STAFF
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"